requirements.txt
streamlit
numpy
Pillow
tensorflow
opencv-python
matplotlib